refactor(ws): reorganize BackgroundTasks and BroadcastConfig structures into separate modules for improved clarity and maintainability (#1128)
