+++ server/websocket/src/broadcast/mod.rs
@@ -1,2 +1,4 @@
 pub mod group;
 pub mod pool;
+pub mod tasks;
+pub mod types;
